Fix process-crashing permission toggle in E2E tests
Reviewer traced a real CI failure (run 28634394559, commit 3eb3f84) to
SetupActivityTest's new setupScreen_showsMediaStep_andAutoAdvancesOnceGranted
test: it crashed the com.gb4pc process outright ("Process crashed"),
aborting the whole connectedDebugAndroidTest shard after only 4 of 21
scheduled tests ran.

Root cause: that test, and every test in PermissionsE2ETest, called
`pm grant` / `pm revoke` for READ_MEDIA_IMAGES against com.gb4pc while
its own instrumented test code was running inside that very process
(`am instrument` attaches to the target app's process, not a separate
one). Changing a storage-group runtime permission on an already-running
process makes Android kill it to re-establish its scoped-storage FUSE
mount, taking down the process the test harness itself depends on.

Fix: never change this permission group on an already-running process.
Remove the crashing SetupActivityTest case entirely (auto-advance
coverage for the MEDIA step already exists at the unit level in
SetupStateTest/PermissionHelperTest). Split PermissionsE2ETest into
PermissionsGrantedE2ETest and PermissionsDeniedE2ETest, neither of
which toggles the permission from inside a test; each assumes a fixed
precondition set before its process starts. connectedE2EAndroidTest
(app/build.gradle.kts) now accepts -PmediaPermissionGranted=false to
`pm revoke` instead of `pm grant` before `am instrument` launches, so
PermissionsDeniedE2ETest gets a dedicated CI step with that flag while
PermissionsGrantedE2ETest keeps the task's existing default-granted
step, both wired into build.yml's summary and failure-gate outcome
computation alongside the other E2E suites.

🦀
